brother

+++ content/data-structures/SplayTree.h
@@ -2,8 +2,8 @@
 struct SplayTree {
 	struct Node {
 		int ch[2] = {0, 0}, p = 0, cnt = 1;
-		long long self = 0, path = 0; // Path aggregates
-		long long sub = 0, vir = 0; // Subtree aggregates
+		ll self = 0, path = 0; // Path aggregates
+		ll sub = 0, vir = 0; // Subtree aggregates
 		bool flip = 0; // Lazy tags
 	};
 	vector<Node> T;
